+++ src/Commands/MakeStateCommand.php
@@ -2,71 +2,45 @@
 
 namespace Erenav\Makers\Commands;
 
+use Illuminate\Support\Str;
 use Symfony\Component\Console\Input\InputArgument;
 
 class MakeStateCommand extends MakersCommand
 {
-    protected $name = 'makers:state';
+    protected $name = 'make:state';
 
-    protected $description = 'Create new state classes';
+    protected $description = 'Create a new state class';
 
-    protected $type = 'States';
-
-    public function handle()
-    {
-        if(parent::handle() === false && !$this->option('force')) {
-            return false;
-        }
-
-        $this->createImplementations();
-    }
-
-    protected function createImplementations(): void
-    {
-        $input = $this->getNameInput();
-
-        $parts = explode('/', strrev($input), 2);
-
-        $parent = strrev($parts[0]);
-
-        $path = strrev($parts[1] ?? '');
-
-        foreach($this->argument('implementations') as $implementation) {
-            $this->call('makers:state-implementation', [
-                'name' => $path . self::NAMESPACE_SEPARATOR . $implementation,
-                'parent' => $parent,
-            ]);
-        }
-    }
+    protected $type = 'State';
 
     protected function getStub(): string
     {
-        return $this->resolveStubPath(config('makers.makes.state.stubs.abstract'));
+        return $this->resolveStubPath(config('makers.makes.state.stubs.standard'));
     }
 
     protected function getDefaultNamespace($rootNamespace): string
     {
-        return $rootNamespace . self::NAMESPACE_SEPARATOR . config('makers.makes.state.namespace');
+        $base = Str::replace('/', self::NAMESPACE_SEPARATOR, $this->argument('base'));
+
+        return $rootNamespace.self::NAMESPACE_SEPARATOR.config('makers.makes.state.namespace').self::NAMESPACE_SEPARATOR.$base;
     }
 
     protected function buildClass($name): string
     {
         $stub = parent::buildClass($name);
 
-        return $this->replaceDefaultState($stub);
+        return $this->replaceParentClass($stub, $this->argument('base'));
     }
 
-    protected function replaceDefaultState(string $stub): string
+    protected function replaceParentClass(string $stub, string $parent): string
     {
-        $default = '';
+        $parts = explode('/', strrev($parent), 2);
 
-        if(!empty($implementations = $this->argument('implementations'))) {
-            $default = '->default(' . $implementations[0] . '::class)';
-        }
+        $parent = strrev($parts[0]);
 
         return str_replace(
-            config('makers.makes.state.merges.default', '{{default}}'),
-            $default,
+            config('makers.makes.state.merges.base', '{{base}}'),
+            $parent,
             $stub
         );
     }
@@ -75,9 +49,9 @@ class MakeStateCommand extends MakersCommand
     {
         return array_merge(parent::getArguments(), [
             [
-                'implementations',
-                InputArgument::OPTIONAL | InputArgument::IS_ARRAY,
-                'The names of the state implementations',
+                'base',
+                InputArgument::REQUIRED,
+                'The name of the base state class',
             ],
         ]);
     }