+++ tests/MakeEnumCommandTest.php
@@ -12,7 +12,7 @@ it('can make an enum', function () {
 
     $name = "{$className}{$postFix}";
 
-    $this->artisan("makers:enum {$className}")->assertExitCode(0);
+    $this->artisan("make:enum {$className}")->assertExitCode(0);
 
     $this->assertFileExists($targetPath . "/{$name}.php");
 });
@@ -30,7 +30,22 @@ it('can make an enum with correct postfix', function () {
 
     $name = "{$className}{$postFix}";
 
-    $this->artisan("makers:enum {$name}")->assertExitCode(0);
+    $this->artisan("make:enum {$name}")->assertExitCode(0);
+
+    $this->assertFileExists($targetPath . "/{$name}.php");
+});
+
+it('can make a backed enum', function () {
+    $targetPath = $this->app->basePath('app/Enums');
+
+    File::deleteDirectory($targetPath);
+
+    $className = 'Test';
+    $postFix = config('makers.makes.enum.postfix');
+
+    $name = "{$className}{$postFix}";
+
+    $this->artisan("make:enum {$className} --type=string")->assertExitCode(0);
 
     $this->assertFileExists($targetPath . "/{$name}.php");
 });