+++ CHANGELOG.md
@@ -5,3 +5,37 @@ All notable changes to `makers` will be documented in this file.
 ## 1.0.0 - 2023-05-10
 
 - Initial Release
+
+
+## 1.0.1 - 2023-05-14
+
+- Update Readme
+
+
+## 1.0.2 - 2023-05-18
+
+- Add make macros command
+
+
+## 1.0.3 - 2023-06-08
+
+- Update Readme
+- Add ability to make value objects
+
+
+## 1.0.4 - 2023-06-08
+
+- Fix value object stub
+
+
+## 1.0.5 - 2023-06-27
+
+- Update make value object command
+
+
+## 2.0.0 - 2024-06-18
+
+- Remove generic factory command
+- Remove state transition command
+- Update command handles
+- Update tests